Add new dependencies; implement reports screen and update product price type to double

+++ lib/features/admin/actions/presentation/screens/admin_product_update_screen.dart
@@ -179,7 +179,7 @@ class _AdminProductUpdateState extends State<AdminProductUpdate> {
                   validator: (value) {
                     if (value?.isEmpty ?? true)
                       return 'Por favor ingrese el precio';
-                    if (int.tryParse(value!) == null)
+                    if (double.tryParse(value!) == null)
                       return 'Ingrese un número válido';
                     return null;
                   },
